Move external PubsubIO hooks outside of PubsubIO.
Also fix these so that they typecheck correctly without relying on java runtime duck typing.
